Old entry trimmer hook.
This hook excludes entries in the import that occurred prior to the last
existing balance entry (that is already in the beancount ledger). Since these
imported transactions should already be accounted for, the output of
bean-extract becomes a lot smaller and easier to read. This is especially useful
for financial institutions that do not allow customizing the date range of
exported transactions.

By default, we only trim entries already marked duplicate (e.g., by
`DuplicateDetector`). If you do not use `DuplicateDetector` or want to trim old
entries regardless of the "duplicate" metadata, set `only_trim_duplicates` to
`False`.